requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
cloudinary==1.35.0
coverage==7.3.0
Django==4.2.4
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.6
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
packaging==23.1
phonenumbers==8.13.24
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
